+++ demo/src/App.tsx
@@ -21,6 +21,7 @@ const App = () => {
     });
 
     const { toolbarPluginInstance } = defaultLayoutPluginInstance;
+    const { searchPluginInstance } = toolbarPluginInstance;
 
     return (
         <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.5.207/build/pdf.worker.js">